--- boa/src/syntax/ast/node/expression/mod.rs
@@ -1,6 +1,0 @@
-//! Expression nodes
-
-pub mod call;
-pub mod new;
-
-pub use self::{call::Call, new::New};
